+++ R/np.plot.engine.conbandwidth.R
@@ -293,6 +293,7 @@
 
       if (quantreg) {
         tobj <- .np_plot_quantile_eval(
+          need.errors = plot.behavior != "plot" || plot.errors.method == "asymptotic",
           txdat = xdat,
           tydat = ydat,
           exdat = x.eval,
@@ -701,6 +702,7 @@
 
         if (quantreg) {
           tobj <- .np_plot_quantile_eval(
+            need.errors = plot.behavior != "plot" || plot.errors.method == "asymptotic",
             txdat = xdat,
             tydat = ydat,
             exdat = subcol(exdat,ei,i)[seq_len(xi.neval),, drop = FALSE],
@@ -994,6 +996,7 @@
 
           if (quantreg) {
             tobj <- .np_plot_quantile_eval(
+              need.errors = plot.behavior != "plot" || plot.errors.method == "asymptotic",
               txdat = xdat,
               tydat = ydat,
               eydat = subcol(eydat,ei,i)[seq_len(xi.neval),, drop = FALSE],